+++ CHANGELOG.md
@@ -29,14 +29,17 @@ Indicates additions to API, minor bug fixes and other small changes.
 - fmt: applied changes from string.h
 - fs: removed fs_directory_delete()
 - job_queue: applied changes from sync.h
+- job_queue: applied changes from thread.h
 - macros/assertions: moved assertion macros from core/macros.h to core/assertions.h
 - macros: added struct_literal() macro for defining struct literals.
 - memory: renamed memory_copy_overlapped to memory_move.
 - path: applied changes from string.h
 - path: rewrote many functions. path_buf_* functions now use AllocatorInterface.
+- platform: applied changes from thread.h
 - platform/win32: applied changes from sync.h
 - platform/win32: applied changes from string.h
 - platform/win32: removed platform_directory_delete()
+- platform/win32: applied changes from thread.h
 - slice: renamed CoreSlice/Buffer to ByteSlice/Buffer
 - slice: renamed GenericSlice/Buffer to ItemSlice/Buffer
 - string: applied changes from slice.h
@@ -46,6 +49,9 @@ Indicates additions to API, minor bug fixes and other small changes.
 - sync: renamed interlocked_* functions to atomic_* and simplified mutex/semaphore api.
 - sync: renamed _fence functions to _barrier
 - test: deleted old tests and replaced with only string tests for now.
+- thread: rewrote thread_create(), thread_destroy() and thread_exit_code()
+- thread: added thread_free(), thread_join_timed() and thread_join()
+- thread: Thread is now struct ThreadHandle
 - types: added 128-bit integers, atomic integers, hash integers, bsize and codepoints.
 - Makefile: bumped version to 0.1.2